+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 
 - Upgraded Android SDK to Level 35
+- Fixed buggy key preview initialisation in landscape mode
 
 
 ## [v1.2.10] (63) Cursed be BroadcastReceiver (2024-05-17)